ODM-88 world type location + attr.AB corrections

Closes ODM-88

See merge request surules/su-rules!48

+++ airline-core-xom/src/ru/integrotech/airline/core/info/PassengerMilesInfo.java
@@ -45,7 +45,7 @@ public class PassengerMilesInfo {
 
 	private String tickedDesignator;
 
-    private double distanceCoeff;
+    private int distanceCoeff;
 
 	private int totalBonusMiles;
 
@@ -93,11 +93,11 @@ public class PassengerMilesInfo {
 		this.airline = airline;
 	}
 
-	public double getDistanceCoeff() {
+	public int getDistanceCoeff() {
 		return distanceCoeff;
 	}
 
-	public void setDistanceCoeff(double distanceCoeff) {
+	public void setDistanceCoeff(int distanceCoeff) {
 		this.distanceCoeff = distanceCoeff;
 	}
